Create llwnn_pso.m
A Local linear Wavelet Neural Network model is advance neural network model which is used to approximate a hypothesis. It has its inherent complexity and is amenable to many changes.

Particle swarm optimization is population based optimization approach. It has been used to optimize the cost function of the LLWNN model keeping in mind the prediction scenario.

Currency Prediction is taken into consideration.